SegundaOpcion/requirements.txt
Flask-OAuthlib
Flask-OAuthlib
Flask-MySQLdb